Feat: Events CRUD

+++ api/adminApi/event/new-api/events-actions.ts
@@ -1,4 +1,4 @@
-import { CRUDBody } from '../../../../interfaces/crud_body_interface';
+import { CRUDBody, RequestType } from '../../../../interfaces/crud_body_interface';
 import asyncHandler from '../../../middlewares/async-handler';
 import { createNewEvent } from './create-new-event';
 import { createNewEvents } from './create-new-events';
@@ -9,34 +9,37 @@ export const eventsActions = asyncHandler(async (request, response) => {
 	if (request.user?.role == 'Admin') {
 		const body: CRUDBody = request.body;
 
-		switch (true) {
-			/**
-			 * Create New Event
-			 **/
-			case body.added.length !== 0 && body.changed.length === 0 && body.deleted.length === 0: {
+		switch (body.requestType) {
+			case RequestType.Create: {
 				const repeatedEvents = body.repeatedEvents;
 
 				if (repeatedEvents.length === 0) {
-					// create 1 event
 					return createNewEvent(request, response, body.pitchId, body.added[0]);
 				} else {
-					// create multiple events
 					return createNewEvents(request, response, body.pitchId, body.added[0], repeatedEvents);
 				}
 			}
 
-			/**
-			 * Update an Event
-			 **/
-			case body.changed.length !== 0 && body.added.length === 0 && body.deleted.length === 0: {
-				return updateEvent(response, body.changed[0]);
-			}
+			case RequestType.Change:
+				{
+					if (body.changed.length && body.added.length === 0)
+						return updateEvent(response, body.changed[0], body.repeatedEvents);
 
-			/**
-			 * Delete an Event
-			 **/
-			case body.deleted.length !== 0 && body.changed.length === 0 && body.added.length === 0:
-				return deleteEvent(response, body.deleted[0]);
+					if (body.added.length && body.changed.length)
+						return updateEvent(
+							response,
+							{ ...body.added[0], seriesId: null, recurrenceException: null, recurrenceRule: null },
+							body.repeatedEvents
+						);
+				}
+				break;
+
+			case RequestType.Remove:
+				{
+					if (body.deleted.length) return deleteEvent(response, body.deleted[0]);
+					if (body.changed.length) return deleteEvent(response, { ...body.changed[0], seriesId: null }, true);
+				}
+				break;
 		}
 	} else {
 		return response.status(401).json({ message: 'Unauthorized' });
